+++ handle-occurrence-filtering-service/src/test/java/io/github/creek/service/ks/connected/services/demo/example/streams/TopologyBuilderTest.java
@@ -16,7 +16,7 @@
 
 package io.github.creek.service.ks.connected.services.demo.example.streams;
 
-import static org.creekservice.api.kafka.metadata.KafkaTopicDescriptor.DEFAULT_CLUSTER_NAME;
+import static org.creekservice.api.kafka.metadata.topic.KafkaTopicDescriptor.DEFAULT_CLUSTER_NAME;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
